Revised Equip System + Flashlight mechanic
I have revised the equip system so that there won't be a need to spawn new objects in order to drop them, instead the same object is used when equipping and on the ground. Also added a flashlight mechanic as I am planning to implement some horror aspects to the game. Made the starting environment dark with some fog.

Device: Laptop