Particle & Sound Effects
Implemented particle and sound effects using Unity's particle pack and the provided free sound sample pack. Added effects only for gun and grenade. Both also have camera shake effects.

The base mechanics for the gun/inventory system is complete, just need to add a variety of guns, heals, grenades and abilities.

Will start work on the level design of the spaceship and the details of the main level.