Spaceship + Different levels
Created a spaceship model with an interior in maya, model textured with SP, then made the ability to move between scenes.
However, with the way my scripts are coded especially for the player and GameManager, it is proven difficult to switch between scenes. I have found a temporary solution by teleporting the player instead of switching to an existing player in the new scene, but it might cause some future problems, I will immediately try to fix this