Improved enemy mechanic + world colliders
I improved the enemy mechanic by using the NavMesh components for better enemy AI instead of the simple MoveTowards() (credits to Fui on youtube) and enemies only detect the player when there are no obstacles between player and enemy ( direct line of sight ). Also added colliders and NavMesh volume filter in the terrain to block the player from unintentionally escaping the terrain, and to save performance by decreasing the area to bake the NavMesh map.