Enemy adjustments, Additional BGM and SFX, Main Menu
I've adjusted the enemy settings and added n tweaked the enemy sounds for better feedback.
I also added more SFX to the player actions and interactions, as well as BGM.
I also made a main menu consisting of Start, Options, Credits, How to Play, and Quit buttons. The content is yet to be filled but i made them working. Will make a save here as I will be making big changes to the project.